Move commands to personal packages
