+++ src/main/kotlin/server/command/grab/GrabArrowCommand.kt
@@ -1,11 +1,13 @@
-package server.command
+package server.command.grab
 
 import game.Game
 import game.player.InventoryItem
 import game.world.RoomContent
+import server.command.Command
 
 class GrabArrowCommand(val game: Game): Command {
     override fun execute() {
+        //TODO clean this up (and the turn commands)
         val playerLocation = game.getPlayerLocation()
         if (game.getWorld().hasRoomContent(playerLocation, RoomContent.ARROW)) {
             game.addToPlayerInventory(InventoryItem.ARROW)
