+++ src/main/kotlin/server/command/grab/GrabFoodCommand.kt
@@ -1,8 +1,9 @@
-package server.command
+package server.command.grab
 
 import game.Game
 import game.player.InventoryItem
 import game.world.RoomContent
+import server.command.Command
 
 class GrabFoodCommand(val game: Game): Command {
     override fun execute() {
