+++ src/main/kotlin/server/command/move/MoveCommand.kt
@@ -1,6 +1,7 @@
-package server.command
+package server.command.move
 
 import game.Game
+import server.command.Command
 import util.Direction
 
 class MoveCommand(private val game: Game, private val direction: Direction): Command {
