+++ src/main/kotlin/server/command/move/MoveEastCommand.kt
@@ -1,6 +1,7 @@
-package server.command
+package server.command.move
 
 import game.Game
+import server.command.Command
 import util.east
 
 class MoveEastCommand(val game: Game): Command {
