+++ src/main/kotlin/server/command/move/MoveNorthCommand.kt
@@ -1,6 +1,7 @@
-package server.command
+package server.command.move
 
 import game.Game
+import server.command.Command
 import util.north
 
 class MoveNorthCommand(val game: Game): Command {
