+++ src/main/kotlin/server/command/move/MoveSouthCommand.kt
@@ -1,6 +1,7 @@
-package server.command
+package server.command.move
 
 import game.Game
+import server.command.Command
 import util.south
 
 class MoveSouthCommand(val game: Game): Command {
