+++ src/main/kotlin/server/command/move/MoveWestCommand.kt
@@ -1,6 +1,7 @@
-package server.command
+package server.command.move
 
 import game.Game
+import server.command.Command
 import util.west
 
 class MoveWestCommand(val game: Game): Command {
