+++ src/test/kotlin/server/command/MoveCommandTest.kt
@@ -8,6 +8,10 @@ import game.player.PlayerState
 import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.MethodSource
+import server.command.move.MoveEastCommand
+import server.command.move.MoveNorthCommand
+import server.command.move.MoveSouthCommand
+import server.command.move.MoveWestCommand
 import util.Direction
 import java.awt.Point
 import java.util.stream.Stream
