+++ src/test/kotlin/server/command/TurnCommandTest.kt
@@ -7,6 +7,8 @@ import game.player.PlayerState
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.MethodSource
+import server.command.turn.TurnLeftCommand
+import server.command.turn.TurnRightCommand
 import util.Direction
 import java.awt.Point
 import java.util.stream.Stream
